fix: 거래 수량 비교를 Decimal로 수정; 정확한 수치 비교를 위해 assert 문 변경

+++ backend/tests/services/test_event_hook.py
@@ -132,4 +132,4 @@ async def test_post_trade_event_hook(monkeypatch):
     assert event["user_id"] == user_id
     assert event["ticker_id"] == ticker_id
     assert event["side"] == side
-    assert float(event["quantity"]) == float(quantity)
+    assert Decimal(event["quantity"]) == quantity
